pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "This project is for property listing and rent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=4.0.2",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "cryptography>=41.0.0",
    "fastapi>=0.117.1",
    "itsdangerous>=2.2.0",
    "pydantic[email]>=2.11.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.37.0",
    "pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "aiohttp>=3.12.15",
    "httpx>=0.28.1",
]
